+++ src/main/resources/static/js/form-builder.js
@@ -601,11 +601,21 @@ var FormBuilder = (function ($) {
         });
     }
 
+    function populateTable(key, data) {
+        var parent = _modal.find('#' + ID_PREFIX.INPUT + key);
+        var tableOpts = findOptionsFormElement(key);
+        parent.empty();
+        parent.append(buildTable(data, tableOpts));
+    }
+
     function buildTable(data, tableOpts) {
-        var columns, columnKeys, headingRow, c, cLen, column, header, key;
-        var style = tableOpts.css || { margin: '0 auto' };
-        var maxHeight = style.hasOwnProperty('max-height') ? style['max-height'] : '450px';
-        var dataTable= $('<table class="form-builder-data-table"/>').css(style);
+        var columns, columnKeys, headingRow, c, cLen, column, header, key, style, maxHeight, dataTable;
+        if (!data) {
+            return;
+        }
+        style = tableOpts.css || { margin: '0 auto' };
+        maxHeight = style.hasOwnProperty('max-height') ? style['max-height'] : '450px';
+        dataTable = $('<table class="form-builder-data-table"/>').css(style);
 
         columns = tableOpts.columns;
         columnKeys = Object.keys(columns);
@@ -847,6 +857,7 @@ var FormBuilder = (function ($) {
                         break;
                     case INPUT_TYPE.TABLE:
                         copyFormTableDataToModel(inputOpts.columns);
+                        break;
                     case INPUT_TYPE.CHECKBOX:
                         val = input.prop('checked');
                         break;
@@ -1093,6 +1104,28 @@ var FormBuilder = (function ($) {
         _modal.find('.modal-body').prepend(warning);
     }
 
+    function findOptionsFormElement(elementKey, formInputs) {
+        var i, len, key, keys, formElement, formElementChild;
+        if (!formInputs) {
+            formInputs = _options.formInputs;
+        }
+        keys = Object.keys(formInputs);
+        for (i = 0, len = keys.length; i < len; i++) {
+            key = keys[i];
+            formElement = formInputs[key];
+            if (key === elementKey) {
+                return formElement;
+            }
+            if (formElement.hasOwnProperty('formInputs')) {
+                formElementChild = findOptionsFormElement(elementKey, formElement.formInputs);
+                if (formElementChild) {
+                    return formElementChild;
+                }
+            }
+        }
+        return null;
+    }
+
     return {
         closeBuilderModal: closeBuilderModal,
         findElementByKey: findElementByKey,
@@ -1102,6 +1135,7 @@ var FormBuilder = (function ($) {
         openBuilderModal: openBuilderModal,
         populateSelect: populateSelect,
         populateTextSelect: populateTextSelect,
+        populateTable: populateTable,
         showAlert: showAlert,
         toggle: toggle,
         setDataValue: setDataValue,