+++ changelog.md
@@ -1,5 +1,5 @@
 ### Revision History
-* 1.1.18-SNAPSHOT
+* 1.1.18
   * Add axis modal now defaults to sorted columns. 
   * Added date pickers to pull request modal.
   * Updated toast to show full message on pull requests errors.
@@ -9,6 +9,7 @@
   * Bug fix: Using the server to get CDN data to prevent CORS issues.
   * Bug fix: Dates sorted correctly for dropdowns on pull request modal.
   * Bug fix: Trim input fields for visualizer find node and cube contains search.
+  * Consumed n-cube 4.1.21
 * 1.1.17
   * Consumed n-cube 4.1.20
 * 1.1.16